feature/applications-module

+++ Modules/Applications/app/Http/Controllers/ApplicationsController/ApplicationController.php
@@ -5,6 +5,7 @@ namespace Modules\Applications\Http\Controllers\ApplicationController;
 use Illuminate\Http\Request;
 use Illuminate\Http\JsonResponse;
 use App\Http\Controllers\Controller;
+use Modules\Applications\Models\Application;
 use Modules\Applications\Services\ApplicationService\ApplicationService;
 use Modules\Applications\Http\Requests\ApplicationRequest\StoreApplicationRequest;
 use Modules\Applications\Http\Requests\ApplicationRequest\UpdateApplicationRequest;
@@ -16,8 +17,9 @@ class ApplicationController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index(): JsonResponse
+    public function index(IndexApplicationRequest $request): JsonResponse
     {
+        $this->authorize('viewAny', Application::class);
         $applications = $this->service->getAllApplications($request->validated());
         return  response()->json($applications);
     }
@@ -27,6 +29,7 @@ class ApplicationController extends Controller
      */
     public function store(StoreApplicationRequest $request): JsonResponse
     {
+        $this->authorize('create', Application::class);
         $application = $this->service->createApplication($request->validated());
         return response()->json($application, 201);
     }
@@ -37,6 +40,7 @@ class ApplicationController extends Controller
     public function show(int $id): JsonResponse
     {
         $application = $this->service->getApplication($id);
+        $this->authorize('view', $application);
         return response()->json($application);
     }
 
@@ -45,6 +49,8 @@ class ApplicationController extends Controller
      */
     public function update(UpdateApplicationRequest $request, int $id): JsonResponse
     {
+        $application = $this->service->getApplication($id);
+        $this->authorize('update', $application);
         $application = $this->service->updateApplication($id, $request->validated());
         return response()->json($application);
     }
@@ -54,14 +60,18 @@ class ApplicationController extends Controller
      */
     public function destroy(int $id): JsonResponse
     {
+        $application = $this->service->getApplication($id);
+        $this->authorize('delete', $application);
         $this->service->deleteApplication($id);
         return response()->json(null, 204);
     }
 
     public function updateStatus(UpdateApplicationRequest $request, int $id): JsonResponse
     {
-        $request->validated();
-        $application = $this->service->updateApplicationStatus($id, $request->status);
+        $application = $this->service->getApplication($id);
+        $this->authorize('update', $application);
+        $request->validate(['status' => 'required|in:pending,approved,rejected']);
+        $application = $this->service->updateApplicationStatus($id, $request);
         return response()->json($application);
-    }
+    } 
 }
