+++ Modules/Applications/app/Http/Controllers/TasksController/TaskController.php
@@ -5,6 +5,7 @@ namespace Modules\Applications\Http\Controllers\TasksController;
 use Illuminate\Http\Request;
 use Illuminate\Http\JsonResponse;
 use App\Http\Controllers\Controller;
+use Modules\Applications\Models\Task;
 use Modules\Applications\Http\Requests\TaskRequest\StoreTaskRequest;
 use Modules\Applications\Http\Requests\TaskRequest\UpdateTaskRequest;
 
@@ -15,8 +16,9 @@ class TaskController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index(): JsonResponse
+    public function index(IndexTaskRequest $request): JsonResponse
     {
+        $this->authorize('viewAny', Task::class);
         $tasks = $this->service->getAllTasks($request->validated());
         return  response()->json($tasks);
     }
@@ -26,6 +28,7 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $request): JsonResponse
     {
+        $this->authorize('create', Task::class);
         $task = $this->service->createTask($request->validated());
         return response()->json($task, 201);
     }
@@ -36,6 +39,7 @@ class TaskController extends Controller
     public function show(int $id): JsonResponse
     {
         $task = $this->service->getTask($id);
+        $this->authorize('view', $task);
         return response()->json($task);
     }
 
@@ -43,7 +47,9 @@ class TaskController extends Controller
      * Update the specified resource in storage.
      */
     public function update(UpdateTaskRequest $request, int $id): JsonResponse
-    {
+    {   
+        $task = $this->service->getTask($id);
+        $this->authorize('update', $task);
         $task = $this->service->updateTask($id, $request->validated());
         return response()->json($task);
     }
@@ -53,14 +59,18 @@ class TaskController extends Controller
      */
     public function destroy(int $id): JsonResponse
     {
+        $task = $this->service->getTask($id);
+        $this->authorize('delete', $task);
         $this->service->deleteTask($id);
         return response()->json(null, 204);
     }
 
-    public function updateStatus(UpdateTaskRequest $request, int $id): JsonResponse
+    public function updateStatus(Request $request, int $id): JsonResponse
     {
-        $request->validated();
+        $task = $this->service->getTask($id);
+        $this->authorize('update', $task);
+        $request->validate(['status' => 'required|in:active,complete']);
         $task = $this->service->updateTaskStatus($id, $request->status);
         return response()->json($task);
-    }
+    } 
 }
